% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-utils.R
\name{check_module_version}
\alias{check_module_version}
\title{Check if a module has a particular version}
\usage{
check_module_version(name, repo = dymiumModulesRepo, version)
}
\arguments{
\item{name}{(\code{character(1)})\cr
A module name.}

\item{repo}{\verb{(character(1)}\cr
A GitHub repository to look for modules.
By default, this uses 'dymium-org/dymiumModules' which is the main repository
that hosts the package and a public repository for shared modules.}

\item{version}{\verb{(character(1)}\cr
Version number of a module. This should follow the following convention,
if you want to download the version 1.1.0 of a module then you should type in
"1.1.0". If this field is left blank, the latest version of the module will be
used.}
}
\value{
a logical value
}
\description{
Check if a module has a particular version
}
\examples{

\dontrun{
  check_module_version('test', version = '0.0.1')
}
}
