% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dymiumModules-package.R
\docType{package}
\name{dymiumModules-package}
\alias{dymiumModules}
\alias{dymiumModules-package}
\title{dymiumModules: a package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
