% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-utils.R
\name{get_modules}
\alias{get_modules}
\title{Get the names of available modules from a remote repository}
\usage{
get_modules(repo = dymiumModulesRepo)
}
\arguments{
\item{repo}{\verb{(character(1)}\cr
A GitHub repository to look for modules.
By default, this uses 'dymium-org/dymiumModules' which is the main repository
that hosts the package and a public repository for shared modules.}

\item{version}{\verb{(character(1)}\cr
Version number of a module. This should follow the following convention,
if you want to download the version 1.1.0 of a module then you should type in
"1.1.0". If this field is left blank, the latest version of the module will be
used.}

\item{name}{(\code{character(1)})\cr
A module name.}
}
\value{
a character vector.
}
\description{
Get the names of available modules from a remote repository
}
\examples{

\dontrun{
  get_modules()
}
}
