% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-utils.R
\name{use_module_readme}
\alias{use_module_readme}
\title{Add a README rmarkdown file to an existing}
\usage{
use_module_readme(name)
}
\arguments{
\item{name}{name of an existing module}
}
\description{
Every module must have a main READMD file which serves as the documentation
of that module. This function provides an easy way to add a README file using
the recommended template.
}
\examples{

\dontrun{
# this assumes that you have a module named 'demography'
use_module_readme(name = "demography")
}
}
