% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-utils.R
\name{get_all_module_files}
\alias{get_all_module_files}
\title{Get all files from all modules in a repository.}
\usage{
get_all_module_files(repo = dymiumModulesRepo)
}
\arguments{
\item{repo}{\verb{(character(1)}\cr
A GitHub repository to look for modules.
By default, this uses 'dymium-org/dymiumModules' which is the main repository
that hosts the package and a public repository for shared modules.}
}
\value{
a character vector.
}
\description{
Get all files from all modules in a repository.
}
\examples{

\dontrun{
  get_all_module_files("dymium-org/dymiumModules")
}
}
