% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-utils.R
\name{get_module_versions}
\alias{get_module_versions}
\title{Get all version numbers of a module}
\usage{
get_module_versions(name, repo = dymiumModulesRepo)
}
\arguments{
\item{name}{(\code{character(1)})\cr
A module name.}

\item{repo}{\verb{(character(1)}\cr
A GitHub repository to look for modules.
By default, this uses 'dymium-org/dymiumModules' which is the main repository
that hosts the package and a public repository for shared modules.}

\item{version}{\verb{(character(1)}\cr
Version number of a module. This should follow the following convention,
if you want to download the version 1.1.0 of a module then you should type in
"1.1.0". If this field is left blank, the latest version of the module will be
used.}
}
\value{
a character vector.
}
\description{
Get all version numbers of a module
}
\examples{

\dontrun{
  get_module_versions("demography")
}
}
